댓글 삭제 POST 처리, 댓글 작성자 이름에 프로필URL 연결, 댓글 작성자에게만 삭제버튼 보이도록 템플릿 분기

+++ reviews/views.py
@@ -1,10 +1,11 @@
-from django.shortcuts import render, redirect
+from django.shortcuts import render, redirect, get_object_or_404
 from reviews.models import Review, Comment
 from .forms import ReviewForm, CommentForm
 from django.contrib.auth.decorators import login_required
 from restaurants.models import Restaurant
 from django.views.generic.list import ListView
-# Create your views here.
+from django.http import HttpResponseForbidden
+
 class ReviewsView(ListView):
     model =Review
     paginate_by = 3
@@ -75,7 +76,6 @@ def update(request, pk):
         # (2) flash message 활용!
         return redirect("reviews:detail", review.pk)
 
-
 @login_required
 def delete(request, pk):
     # pk에 해당하는 글 삭제
@@ -83,11 +83,9 @@ def delete(request, pk):
     review.delete()
     return redirect("reviews:index")
 
-
-
 @login_required
 def comment_create(request, pk):
-    review = Review.objects.get(pk=pk)
+    review = get_object_or_404(Review, pk=pk)
     comment_form = CommentForm(request.POST)
     if comment_form.is_valid():
         comment = comment_form.save(commit=False)
@@ -95,10 +93,13 @@ def comment_create(request, pk):
         comment.user = request.user
         comment.save()
     return redirect("reviews:detail", review.pk)
+
 @login_required
 def comment_delete(request, pk, comment_pk):
-    comment = Comment.objects.get(pk=comment_pk)
-
+    comment = get_object_or_404(Comment, pk=comment_pk)
     if comment.user == request.user:
-        comment.delete()
-    return redirect("reviews:detail", pk)
+        if request.method == 'POST':
+            comment.delete()
+            return redirect("reviews:detail", pk)
+    else:
+        return HttpResponseForbidden()
